+++ Commands/Admin/NewTenantSite.cs
@@ -13,7 +13,7 @@ using Resources = SharePointPnP.PowerShell.Commands.Properties.Resources;
 namespace SharePointPnP.PowerShell.Commands
 {
     [Cmdlet(VerbsCommon.New, "PnPTenantSite")]
-    [CmdletHelp("Creates a new site collection for the current tenant",
+    [CmdletHelp("Creates a new (classic) site collection for the current tenant",
         @"The New-PnPTenantSite cmdlet creates a new site collection for the current company. However, creating a new SharePoint
 Online site collection fails if a deleted site with the same URL exists in the Recycle Bin. If you want to use this command for an on-premises farm, please refer to http://blogs.msdn.com/b/vesku/archive/2014/06/09/provisioning-site-collections-using-sp-app-model-in-on-premises-with-just-csom.aspx ",
         Category = CmdletHelpCategory.TenantAdmin)]