+++ Commands/Base/GetContext.cs
@@ -15,6 +15,14 @@ namespace SharePointPnP.PowerShell.Commands.Base
         Code = @"PS:> $ctx = Get-PnPContext",
         Remarks = @"This will put the current context in the $ctx variable.",
         SortOrder = 1)]
+            [CmdletExample(
+        Code = @"PS:> Connect-PnPOnline -Url $siteAurl -Credentials $credentials
+PS:> $ctx = Get-PnPContext
+PS:> Get-PnPList # returns the lists from site specified with $siteAurl
+PS:> Connect-PnPOnline -Url $siteBurl -Credentials $credentials
+PS:> Get-PnPList # returns the lists from the site specified with $siteBurl
+PS:> Set-PnPContext -Context $ctx # switch back to site A
+PS:> Get-PnPList # returns the lists from site A", SortOrder = 2)]
     public class GetSPOContext : PSCmdlet
     {