+++ Commands/Lists/GetListItem.cs
@@ -76,6 +76,8 @@ namespace SharePointPnP.PowerShell.Commands.Lists
 		protected override void ExecuteCmdlet()
         {
             var list = List.GetList(SelectedWeb);
+            if (list == null)
+                throw new PSArgumentException($"No list found with id, title or url '{List}'", "List");
 
             if (HasId())
             {